ref(optimizations): move to correct package and correct mixins config (forgot to push last one)
