+++ arclight-common/src/main/java/io/izzel/arclight/common/mixin/core/entity/EntityMixin.java
@@ -84,11 +84,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 import javax.annotation.Nullable;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
-import java.util.Random;
-import java.util.UUID;
+import java.util.*;
 
 @SuppressWarnings("ConstantConditions")
 @Mixin(Entity.class)
@@ -598,6 +594,14 @@ public abstract class EntityMixin implements InternalEntityBridge, EntityBridge,
     public List<Entity> bridge$getPassengers() {
         return passengers;
     }
+
+    @Override
+    public void bridge$getRecursivePassengers(boolean playersOnly, Set<Entity> passengers) {
+        this.getRecursivePassengers(playersOnly, passengers);
+    }
+
+    @Shadow
+    protected abstract void getRecursivePassengers(boolean playersOnly, Set<Entity> passengers);
 
     @Override
     public boolean bridge$addPassenger(Entity entity) {
