+++ arclight-common/src/main/java/io/izzel/arclight/common/mixin/core/world/TrackedEntityMixin.java
@@ -100,6 +100,11 @@ public abstract class TrackedEntityMixin implements TrackedEntityBridge {
         this.trackedPlayers = trackedPlayers;
     }
 
+    @Override
+    public void bridge$setPassengers(List<Entity> passengers) {
+        this.passengers = passengers;
+    }
+
     /**
      * @author IzzelAliz
      * @reason
