+++ arclight-forge/src/main/resources/mixins.arclight.impl.forge.optimization.stream.json
@@ -6,6 +6,14 @@
   "mixins": [
     "CommandNodeMixin_Optimization",
     "GoalSelectorMixin",
-    "ITaskQueue_PriorityMixin"
+    "ITaskQueue_PriorityMixin",
+    "AbstractHorseEntityMixin_Optimize",
+    "AbstractMinecraftEntityMixin_Optimize",
+    "BoatEntityMixin_Optimize",
+    "EntityMixin_Optimize",
+    "MixinTrackedEntity_Optimize",
+    "PigEntityMixin_Optimize",
+    "RavagerEntityMixin_Optimize",
+    "SSetPassengersPacketMixin_Optimize"
   ]
 }
